+++ client/src/components/table/cell.tsx
@@ -5,7 +5,6 @@ import {
   SortDescendingIcon,
 } from '@heroicons/react/solid';
 import classNames from 'classnames';
-import { useMemo } from 'react';
 
 import type { CellContext, HeaderContext } from '@tanstack/react-table';
 
@@ -29,34 +28,37 @@ const CellWrapper = <T, C>({ children, context }: React.PropsWithChildren<CellPr
   const isFirstColumn = context.table.getAllColumns()[0].id === context.column.id;
 
   const { align } = context.column.columnDef?.meta || {};
-  const style = useMemo(
-    () => ({
-      paddingLeft: isFirstColumn ? `${context.row.depth * 20 + 25}px` : '25px',
-    }),
-    [context.row.depth, isFirstColumn],
-  );
+  const { depth } = context.row;
 
   const isExpandible = isFirstColumn && context.row.getCanExpand();
   const isExpanded = isExpandible && context.row.getIsExpanded();
   const toggleExpand = context.row.getToggleExpandedHandler();
+  const canExpand = context.table.getCanSomeRowsExpand();
 
   return (
     <div
       onClick={isExpandible ? toggleExpand : undefined}
-      style={style}
       className={classNames(
         getAlignmentClasses(align),
-        'pr-5 relative flex items-center justify-start w-full h-20 text-sm',
-        { 'cursor-pointer': isExpandible },
+        'relative flex items-center justify-start w-full min-h-20',
+        {
+          'cursor-pointer': isExpandible,
+          'pl-2.5 font-bold text-sm': !canExpand && isFirstColumn && depth === 0,
+          'pl-4 font-normal text-sm': canExpand && isFirstColumn && depth === 0,
+          'pl-10 uppercase text-2xs font-semibold': isFirstColumn && depth === 1,
+          'pl-[72px] text-xs font-normal': isFirstColumn && depth === 2,
+          'text-sm': !isFirstColumn,
+        },
       )}
     >
       <div className="w-full mx-auto my-auto">
-        {isExpandible ? (
-          <div>
-            <div className="absolute -translate-x-5 -translate-y-1/2 top-1/2">
+        {canExpand ? (
+          <div className="flex items-center gap-4">
+            <div className="w-4 h-4">
               <ChevronRightIcon
-                className={classNames('w-4 h-4 text-gray-900', {
+                className={classNames(' text-gray-900', {
                   'rotate-90': isExpanded,
+                  hidden: !isExpandible,
                 })}
               />
             </div>
@@ -90,10 +92,9 @@ export const HeaderCell = <T, C>({
       })}
     >
       <div
-        className={classNames(
-          'flex flex-row justify-center items-center gap-2',
-          getAlignmentClasses(align),
-        )}
+        className={classNames('flex flex-row gap-2 items-center', getAlignmentClasses(align), {
+          'justify-center': align === 'center',
+        })}
       >
         <div>{children}</div>
         {context.column.getCanSort() && (